refactor(card-browser): move Saved Searches to Fragment
* move onSaveSearchNamePrompt to Fragment
