+++ AnkiDroid/src/main/java/com/ichi2/anki/CardBrowser.kt
@@ -66,7 +66,6 @@ import com.ichi2.anki.common.annotations.NeedsTest
 import com.ichi2.anki.common.utils.annotation.KotlinCleanup
 import com.ichi2.anki.databinding.ActivityCardBrowserBinding
 import com.ichi2.anki.dialogs.DiscardChangesDialog
-import com.ichi2.anki.dialogs.SaveBrowserSearchDialogFragment
 import com.ichi2.anki.dialogs.SavedBrowserSearchesDialogFragment
 import com.ichi2.anki.dialogs.registerDeckSelectedHandler
 import com.ichi2.anki.dialogs.registerSaveSearchHandler
@@ -90,7 +89,6 @@ import com.ichi2.anki.ui.ResizablePaneManager
 import com.ichi2.anki.utils.ext.addPrepareMenuProvider
 import com.ichi2.anki.utils.ext.launchCollectionInLifecycleScope
 import com.ichi2.anki.utils.ext.onAllFragmentsLoaded
-import com.ichi2.anki.utils.ext.showDialogFragment
 import com.ichi2.ui.CardBrowserSearchView
 import com.ichi2.utils.AndroidUiUtils.hideKeyboard
 import com.ichi2.utils.LanguageUtil
@@ -529,13 +527,6 @@ open class CardBrowser :
             }
         }
 
-        fun onSaveSearchNamePrompt(searchTerms: String) {
-            Timber.i("opening 'save search' name input dialog")
-            val dialog =
-                SaveBrowserSearchDialogFragment.newInstance(searchQuery = searchTerms)
-            showDialogFragment(dialog)
-        }
-
         viewModel.flowOfSearchQueryExpanded.launchCollectionInLifecycleScope(::onSearchQueryExpanded)
         viewModel.flowOfSelectedRows.launchCollectionInLifecycleScope(::onSelectedRowsChanged)
         viewModel.flowOfFilterQuery.launchCollectionInLifecycleScope(::onFilterQueryChanged)
@@ -543,7 +534,6 @@ open class CardBrowser :
         viewModel.flowOfMultiSelectModeChanged.launchCollectionInLifecycleScope(::onMultiSelectModeChanged)
         viewModel.flowOfSearchState.launchCollectionInLifecycleScope(::searchStateChanged)
         viewModel.flowOfNoteEditorCommand.launchCollectionInLifecycleScope(::onNoteEditorCommand)
-        viewModel.flowOfSaveSearchNamePrompt.launchCollectionInLifecycleScope(::onSaveSearchNamePrompt)
     }
 
     private fun hideKeyboard() {
@@ -616,17 +606,6 @@ open class CardBrowser :
                     return true
                 }
             }
-            KeyEvent.KEYCODE_S -> {
-                if (event.isCtrlPressed && event.isAltPressed) {
-                    Timber.i("Ctrl+Alt+S: Show saved searches")
-                    showSavedSearches()
-                    return true
-                } else if (event.isCtrlPressed) {
-                    Timber.i("Ctrl+S: Save search")
-                    viewModel.saveCurrentSearch()
-                    return true
-                }
-            }
             in KeyEvent.KEYCODE_1..KeyEvent.KEYCODE_7 -> {
                 if (event.isCtrlPressed) {
                     Timber.i("Update flag")
@@ -684,16 +663,6 @@ open class CardBrowser :
             viewModel.endMultiSelectMode(SingleSelectCause.NavigateBack)
         } else {
             super.onNavigationPressed()
-        }
-    }
-
-    fun showSavedSearches() {
-        launchCatchingTask {
-            val dialog =
-                SavedBrowserSearchesDialogFragment.newInstance(
-                    viewModel.savedSearches(),
-                )
-            showDialogFragment(dialog)
         }
     }
